requirements.txt
google-generativeai
pillow
streamlit
loguru
python-dotenv